+++ comnetsemu/test/unit/test_net.py
@@ -81,14 +81,15 @@ class TestVNFManager(unittest.TestCase):
                     docker_args={"cpu_quota": 1000},
                 )
                 cname_list.append(f"c{i}{j}")
+        cname_list_get = self.mgr.getAllContainers()
+        self.assertEqual(cname_list, cname_list_get)
 
         # Check docker_args works
         c11_ins = self.mgr._getDockerIns("c11")
         self.assertEqual(c11_ins.attrs["HostConfig"]["CpuQuota"], 1000)
 
         for i in range(1, HOST_NUM + 1):
-            cins_host = self.mgr.getContainers(f"h{i}")
-            cname_list_host = [c.name for c in cins_host]
+            cname_list_host = self.mgr.getContainersDhost(f"h{i}")
             self.assertEqual(cname_list_host, [f"c{i}{k}" for k in range(1, i + 1)])
             for cname in cname_list_host:
                 self.mgr.removeContainer(cname)
@@ -96,6 +97,13 @@ class TestVNFManager(unittest.TestCase):
         for cname in cname_list:
             c_ins = self.mgr._getDockerIns(cname)
             self.assertTrue(c_ins is None)
+
+        d1 = self.mgr.addContainer("d1", "h1", "dev_test", "bash", docker_args={})
+        d1_get = self.mgr.getContainerInstance("d1")
+        self.assertIs(d1, d1_get)
+        self.mgr.removeContainer("d1")
+        d2 = self.mgr.getContainerInstance("d2", None)
+        self.assertEqual(d2, None)
 
     @unittest.skipIf(len(sys.argv) == 2 and sys.argv[1] == "-f", "Schneller!")
     def test_container_isolation(self):
@@ -110,7 +118,7 @@ class TestVNFManager(unittest.TestCase):
         c1 = self.mgr.addContainer(
             "c1", "h1", "dev_test", "stress-ng -c 1 -m 1 --vm-bytes 300M", {}
         )
-        usages = self.mgr.monResourceStats(c1.name, sample_period=0.1)
+        usages = self.mgr.monResourceStats(c1.name, sample_period=0.1, sample_num=2)
         cpu = sum(u[0] for u in usages) / len(usages)
         mem = sum(u[1] for u in usages) / len(usages)
         self.assertTrue(abs(cpu - 10.0) <= CPU_ERR_THR)
@@ -133,6 +141,9 @@ class TestVNFManager(unittest.TestCase):
     #     from comnetsemu.cli import CLI
     #     CLI(self.net)
     #     self.mgr.removeContainer("c1")
+
+    def test_appcontainermanager_rest_api(self):
+        pass
 
 
 if __name__ == "__main__":